refactor: 정적파일 assets 폴더로 이동
