+++ src/assets/landing/section1-2.png
