+++ src/assets/landing/section2-1.png
