+++ src/assets/landing/section2-2.png
