+++ src/assets/landing/section3-1.png
