+++ src/assets/landing/section3-2.png
