+++ src/page/LandingPage/components/MainSection.tsx
@@ -1,5 +1,4 @@
-// src/components/landing/MainSection/MainSection.tsx
-import preview from '../../../../public/landing/preview.webm';
+import preview from '../../../assets/landing/preview.webm';
 
 interface MainSectionProps {
   onStartWithoutLogin: () => void;
@@ -11,7 +10,9 @@ export default function MainSection({ onStartWithoutLogin }: MainSectionProps) {
       id="main-section"
       className="flex flex-col items-center justify-center gap-12"
     >
-      <video src={preview} autoPlay muted loop className="w-2/3" />
+      <video src={preview} autoPlay muted loop className="w-2/3">
+        <p>브라우저에서 비디오를 지원하지 않습니다.</p>
+      </video>
       <h1 className="text-[3.5vw] font-bold">토론 진행을 더 쉽고 빠르게</h1>
       <button
         onClick={onStartWithoutLogin}
